pyproject.toml
[project]
name = "repl-nix-bettersystemsaiweb"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
]
